refactor(query): tests from query/tests to query/src
