+++ src/query/src/lib.rs
@@ -26,6 +26,8 @@ pub mod plan;
 pub mod planner;
 pub mod query_engine;
 pub mod sql;
+#[cfg(test)]
+mod tests;
 
 pub use crate::datafusion::DfContextProviderAdapter;
 pub use crate::query_engine::{
