+++ src/query/src/tests.rs
@@ -1,0 +1,14 @@
+
+mod argmax_test;
+mod argmin_test;
+mod mean_test;
+mod my_sum_udaf_example;
+mod percentile_test;
+mod polyval_test;
+mod query_engine_test;
+mod scipy_stats_norm_cdf_test;
+mod scipy_stats_norm_pdf;
+mod time_range_filter_test;
+
+mod pow;
+mod function;
