+++ src/query/src/tests/argmin_test.rs
@@ -12,7 +12,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-mod function;
 
 use std::sync::Arc;
 
@@ -22,64 +21,71 @@ use common_recordbatch::{util, RecordBatch};
 use datatypes::for_all_primitive_types;
 use datatypes::prelude::*;
 use datatypes::types::WrapperType;
-use datatypes::value::OrderedFloat;
-use format_num::NumberFormat;
-use num_traits::AsPrimitive;
-use query::error::Result;
-use query::parser::QueryLanguageParser;
-use query::QueryEngine;
 use session::context::QueryContext;
 
+use crate::error::Result;
+use crate::parser::QueryLanguageParser;
+use crate::QueryEngine;
+use crate::tests::function;
+
+
 #[tokio::test]
-async fn test_mean_aggregator() -> Result<()> {
+async fn test_argmin_aggregator() -> Result<()> {
     common_telemetry::init_default_ut_logging();
     let engine = function::create_query_engine();
 
-    macro_rules! test_mean {
+    macro_rules! test_argmin {
         ([], $( { $T:ty } ),*) => {
             $(
                 let column_name = format!("{}_number", std::any::type_name::<$T>());
-                test_mean_success::<$T>(&column_name, "numbers", engine.clone()).await?;
+                test_argmin_success::<$T>(&column_name, "numbers", engine.clone()).await?;
             )*
         }
     }
-    for_all_primitive_types! { test_mean }
+    for_all_primitive_types! { test_argmin }
     Ok(())
 }
 
-async fn test_mean_success<T>(
+async fn test_argmin_success<T>(
     column_name: &str,
     table_name: &str,
     engine: Arc<dyn QueryEngine>,
 ) -> Result<()>
 where
-    T: WrapperType + AsPrimitive<f64>,
+    T: WrapperType + PartialOrd,
 {
-    let result = execute_mean(column_name, table_name, engine.clone())
+    let result = execute_argmin(column_name, table_name, engine.clone())
         .await
         .unwrap();
-    let value = function::get_value_from_batches("mean", result);
+    let value = function::get_value_from_batches("argmin", result);
 
     let numbers =
         function::get_numbers_from_table::<T>(column_name, table_name, engine.clone()).await;
-    let expected_value = numbers.iter().map(|&n| n.as_()).collect::<Vec<f64>>();
-
-    let expected_value = inc_stats::mean(expected_value.iter().cloned()).unwrap();
-    if let Value::Float64(OrderedFloat(value)) = value {
-        let num = NumberFormat::new();
-        let value = num.format(".6e", value);
-        let expected_value = num.format(".6e", expected_value);
-        assert_eq!(value, expected_value);
-    }
+    let expected_value = match numbers.len() {
+        0 => 0_u32,
+        _ => {
+            let mut index = 0;
+            let mut min = numbers[0];
+            for (i, &number) in numbers.iter().enumerate() {
+                if min > number {
+                    min = number;
+                    index = i;
+                }
+            }
+            index as u32
+        }
+    };
+    let expected_value = Value::from(expected_value);
+    assert_eq!(value, expected_value);
     Ok(())
 }
 
-async fn execute_mean<'a>(
+async fn execute_argmin<'a>(
     column_name: &'a str,
     table_name: &'a str,
     engine: Arc<dyn QueryEngine>,
 ) -> RecordResult<Vec<RecordBatch>> {
-    let sql = format!("select MEAN({column_name}) as mean from {table_name}");
+    let sql = format!("select argmin({column_name}) as argmin from {table_name}");
     let stmt = QueryLanguageParser::parse_sql(&sql).unwrap();
     let plan = engine
         .statement_to_plan(stmt, Arc::new(QueryContext::new()))
