+++ src/query/src/tests/mean_test.rs
@@ -12,8 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+
 use std::sync::Arc;
-mod function;
 
 use common_query::Output;
 use common_recordbatch::error::Result as RecordResult;
@@ -21,65 +21,66 @@ use common_recordbatch::{util, RecordBatch};
 use datatypes::for_all_primitive_types;
 use datatypes::prelude::*;
 use datatypes::types::WrapperType;
+use datatypes::value::OrderedFloat;
+use format_num::NumberFormat;
 use num_traits::AsPrimitive;
-use query::error::Result;
-use query::parser::QueryLanguageParser;
-use query::QueryEngine;
 use session::context::QueryContext;
 
+use crate::error::Result;
+use crate::parser::QueryLanguageParser;
+use crate::QueryEngine;
+use crate::tests::function;
+
 #[tokio::test]
-async fn test_polyval_aggregator() -> Result<()> {
+async fn test_mean_aggregator() -> Result<()> {
     common_telemetry::init_default_ut_logging();
     let engine = function::create_query_engine();
 
-    macro_rules! test_polyval {
+    macro_rules! test_mean {
         ([], $( { $T:ty } ),*) => {
             $(
                 let column_name = format!("{}_number", std::any::type_name::<$T>());
-                test_polyval_success::<$T, <<<$T as WrapperType>::LogicalType as LogicalPrimitiveType>::LargestType as LogicalPrimitiveType>::Wrapper>(&column_name, "numbers", engine.clone()).await?;
+                test_mean_success::<$T>(&column_name, "numbers", engine.clone()).await?;
             )*
         }
     }
-    for_all_primitive_types! { test_polyval }
+    for_all_primitive_types! { test_mean }
     Ok(())
 }
 
-async fn test_polyval_success<T, PolyT>(
+async fn test_mean_success<T>(
     column_name: &str,
     table_name: &str,
     engine: Arc<dyn QueryEngine>,
 ) -> Result<()>
 where
-    T: WrapperType,
-    PolyT: WrapperType,
-    T::Native: AsPrimitive<PolyT::Native>,
-    PolyT::Native: std::ops::Mul<Output = PolyT::Native> + std::iter::Sum,
-    i64: AsPrimitive<PolyT::Native>,
+    T: WrapperType + AsPrimitive<f64>,
 {
-    let result = execute_polyval(column_name, table_name, engine.clone())
+    let result = execute_mean(column_name, table_name, engine.clone())
         .await
         .unwrap();
-    let value = function::get_value_from_batches("polyval", result);
+    let value = function::get_value_from_batches("mean", result);
 
     let numbers =
         function::get_numbers_from_table::<T>(column_name, table_name, engine.clone()).await;
-    let expected_value = numbers.iter().copied();
-    let x = 0i64;
-    let len = expected_value.len();
-    let expected_native: PolyT::Native = expected_value
-        .enumerate()
-        .map(|(i, v)| v.into_native().as_() * (x.pow((len - 1 - i) as u32)).as_())
-        .sum();
-    assert_eq!(value, PolyT::from_native(expected_native).into());
+    let expected_value = numbers.iter().map(|&n| n.as_()).collect::<Vec<f64>>();
+
+    let expected_value = inc_stats::mean(expected_value.iter().cloned()).unwrap();
+    if let Value::Float64(OrderedFloat(value)) = value {
+        let num = NumberFormat::new();
+        let value = num.format(".6e", value);
+        let expected_value = num.format(".6e", expected_value);
+        assert_eq!(value, expected_value);
+    }
     Ok(())
 }
 
-async fn execute_polyval<'a>(
+async fn execute_mean<'a>(
     column_name: &'a str,
     table_name: &'a str,
     engine: Arc<dyn QueryEngine>,
 ) -> RecordResult<Vec<RecordBatch>> {
-    let sql = format!("select POLYVAL({column_name}, 0) as polyval from {table_name}");
+    let sql = format!("select MEAN({column_name}) as mean from {table_name}");
     let stmt = QueryLanguageParser::parse_sql(&sql).unwrap();
     let plan = engine
         .statement_to_plan(stmt, Arc::new(QueryContext::new()))
