+++ src/query/src/tests/percentile_test.rs
@@ -13,7 +13,6 @@
 // limitations under the License.
 
 use std::sync::Arc;
-mod function;
 use catalog::local::{MemoryCatalogManager, MemoryCatalogProvider, MemorySchemaProvider};
 use catalog::{CatalogList, CatalogProvider, SchemaProvider};
 use common_catalog::consts::{DEFAULT_CATALOG_NAME, DEFAULT_SCHEMA_NAME};
@@ -26,11 +25,15 @@ use datatypes::schema::{ColumnSchema, Schema};
 use datatypes::vectors::Int32Vector;
 use function::{create_query_engine, get_numbers_from_table};
 use num_traits::AsPrimitive;
-use query::error::Result;
-use query::parser::QueryLanguageParser;
-use query::{QueryEngine, QueryEngineFactory};
+
 use session::context::QueryContext;
 use table::test_util::MemTable;
+
+use crate::error::Result;
+use crate::{QueryEngine, QueryEngineFactory};
+use crate::parser::QueryLanguageParser;
+use crate::tests::function;
+
 
 #[tokio::test]
 async fn test_percentile_aggregator() -> Result<()> {
