+++ src/query/src/tests/polyval_test.rs
@@ -12,8 +12,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-mod function;
-
 use std::sync::Arc;
 
 use common_query::Output;
@@ -22,68 +20,67 @@ use common_recordbatch::{util, RecordBatch};
 use datatypes::for_all_primitive_types;
 use datatypes::prelude::*;
 use datatypes::types::WrapperType;
-use query::error::Result;
-use query::parser::QueryLanguageParser;
-use query::QueryEngine;
+use num_traits::AsPrimitive;
 use session::context::QueryContext;
 
+use crate::error::Result;
+use crate::parser::QueryLanguageParser;
+use crate::QueryEngine;
+use crate::tests::function;
+
 #[tokio::test]
-async fn test_argmax_aggregator() -> Result<()> {
+async fn test_polyval_aggregator() -> Result<()> {
     common_telemetry::init_default_ut_logging();
     let engine = function::create_query_engine();
 
-    macro_rules! test_argmax {
+    macro_rules! test_polyval {
         ([], $( { $T:ty } ),*) => {
             $(
                 let column_name = format!("{}_number", std::any::type_name::<$T>());
-                test_argmax_success::<$T>(&column_name, "numbers", engine.clone()).await?;
+                test_polyval_success::<$T, <<<$T as WrapperType>::LogicalType as LogicalPrimitiveType>::LargestType as LogicalPrimitiveType>::Wrapper>(&column_name, "numbers", engine.clone()).await?;
             )*
         }
     }
-    for_all_primitive_types! { test_argmax }
+    for_all_primitive_types! { test_polyval }
     Ok(())
 }
 
-async fn test_argmax_success<T>(
+async fn test_polyval_success<T, PolyT>(
     column_name: &str,
     table_name: &str,
     engine: Arc<dyn QueryEngine>,
 ) -> Result<()>
 where
-    T: WrapperType + PartialOrd,
+    T: WrapperType,
+    PolyT: WrapperType,
+    T::Native: AsPrimitive<PolyT::Native>,
+    PolyT::Native: std::ops::Mul<Output = PolyT::Native> + std::iter::Sum,
+    i64: AsPrimitive<PolyT::Native>,
 {
-    let result = execute_argmax(column_name, table_name, engine.clone())
+    let result = execute_polyval(column_name, table_name, engine.clone())
         .await
         .unwrap();
-    let value = function::get_value_from_batches("argmax", result);
+    let value = function::get_value_from_batches("polyval", result);
 
     let numbers =
         function::get_numbers_from_table::<T>(column_name, table_name, engine.clone()).await;
-    let expected_value = match numbers.len() {
-        0 => 0_u64,
-        _ => {
-            let mut index = 0;
-            let mut max = numbers[0];
-            for (i, &number) in numbers.iter().enumerate() {
-                if max < number {
-                    max = number;
-                    index = i;
-                }
-            }
-            index as u64
-        }
-    };
-    let expected_value = Value::from(expected_value);
-    assert_eq!(value, expected_value);
+    let expected_value = numbers.iter().copied();
+    let x = 0i64;
+    let len = expected_value.len();
+    let expected_native: PolyT::Native = expected_value
+        .enumerate()
+        .map(|(i, v)| v.into_native().as_() * (x.pow((len - 1 - i) as u32)).as_())
+        .sum();
+    assert_eq!(value, PolyT::from_native(expected_native).into());
     Ok(())
 }
 
-async fn execute_argmax<'a>(
+async fn execute_polyval<'a>(
     column_name: &'a str,
     table_name: &'a str,
     engine: Arc<dyn QueryEngine>,
 ) -> RecordResult<Vec<RecordBatch>> {
-    let sql = format!("select ARGMAX({column_name}) as argmax from {table_name}");
+    let sql = format!("select POLYVAL({column_name}, 0) as polyval from {table_name}");
     let stmt = QueryLanguageParser::parse_sql(&sql).unwrap();
     let plan = engine
         .statement_to_plan(stmt, Arc::new(QueryContext::new()))
