+++ src/query/src/tests/scipy_stats_norm_pdf.rs
@@ -13,7 +13,6 @@
 // limitations under the License.
 
 use std::sync::Arc;
-mod function;
 
 use common_query::Output;
 use common_recordbatch::error::Result as RecordResult;
@@ -21,31 +20,34 @@ use common_recordbatch::{util, RecordBatch};
 use datatypes::for_all_primitive_types;
 use datatypes::types::WrapperType;
 use num_traits::AsPrimitive;
-use query::error::Result;
-use query::parser::QueryLanguageParser;
-use query::QueryEngine;
 use session::context::QueryContext;
-use statrs::distribution::{ContinuousCDF, Normal};
+use statrs::distribution::{Continuous, Normal};
 use statrs::statistics::Statistics;
 
+use crate::error::Result;
+use crate::parser::QueryLanguageParser;
+use crate::QueryEngine;
+use crate::tests::function;
+
+
 #[tokio::test]
-async fn test_scipy_stats_norm_cdf_aggregator() -> Result<()> {
+async fn test_scipy_stats_norm_pdf_aggregator() -> Result<()> {
     common_telemetry::init_default_ut_logging();
     let engine = function::create_query_engine();
 
-    macro_rules! test_scipy_stats_norm_cdf {
+    macro_rules! test_scipy_stats_norm_pdf {
         ([], $( { $T:ty } ),*) => {
             $(
                 let column_name = format!("{}_number", std::any::type_name::<$T>());
-                test_scipy_stats_norm_cdf_success::<$T>(&column_name, "numbers", engine.clone()).await?;
+                test_scipy_stats_norm_pdf_success::<$T>(&column_name, "numbers", engine.clone()).await?;
             )*
         }
     }
-    for_all_primitive_types! { test_scipy_stats_norm_cdf }
+    for_all_primitive_types! { test_scipy_stats_norm_pdf }
     Ok(())
 }
 
-async fn test_scipy_stats_norm_cdf_success<T>(
+async fn test_scipy_stats_norm_pdf_success<T>(
     column_name: &str,
     table_name: &str,
     engine: Arc<dyn QueryEngine>,
@@ -53,10 +55,10 @@ async fn test_scipy_stats_norm_cdf_success<T>(
 where
     T: WrapperType + AsPrimitive<f64>,
 {
-    let result = execute_scipy_stats_norm_cdf(column_name, table_name, engine.clone())
+    let result = execute_scipy_stats_norm_pdf(column_name, table_name, engine.clone())
         .await
         .unwrap();
-    let value = function::get_value_from_batches("scipy_stats_norm_cdf", result);
+    let value = function::get_value_from_batches("scipy_stats_norm_pdf", result);
 
     let numbers =
         function::get_numbers_from_table::<T>(column_name, table_name, engine.clone()).await;
@@ -65,19 +67,19 @@ where
     let stddev = expected_value.std_dev();
 
     let n = Normal::new(mean, stddev).unwrap();
-    let expected_value = n.cdf(2.0);
+    let expected_value = n.pdf(2.0);
 
     assert_eq!(value, expected_value.into());
     Ok(())
 }
 
-async fn execute_scipy_stats_norm_cdf<'a>(
+async fn execute_scipy_stats_norm_pdf<'a>(
     column_name: &'a str,
     table_name: &'a str,
     engine: Arc<dyn QueryEngine>,
 ) -> RecordResult<Vec<RecordBatch>> {
     let sql = format!(
-        "select SCIPYSTATSNORMCDF({column_name},2.0) as scipy_stats_norm_cdf from {table_name}",
+        "select SCIPYSTATSNORMPDF({column_name},2.0) as scipy_stats_norm_pdf from {table_name}"
     );
     let stmt = QueryLanguageParser::parse_sql(&sql).unwrap();
     let plan = engine
